Move the unpub_aws package to the root of the project, so it can be released separately from the main unpub package. Add json_annotations to the unpub_api to stop errors in compiler.
