+++ unpub/lib/unpub_api/pubspec.yaml
@@ -2,3 +2,6 @@ name: unpub_api
 
 environment:
   sdk: ">=2.12.0 <3.0.0"
+
+dependencies:
+  json_annotation: ^4.1.0
