--- unpub/lib/unpub_aws/unpub_aws.dart
@@ -1,2 +1,0 @@
-export 'package:unpub/unpub_aws/lib/s3_file_store.dart';
-export 'package:unpub/unpub_aws/lib/aws_credentials.dart';
