+++ unpub_aws/lib/unpub_aws.dart
@@ -1,0 +1,2 @@
+export 's3/s3_file_store.dart';
+export 'core/aws_credentials.dart';
