+++ unpub_aws/test/s3_store_test.dart
@@ -3,8 +3,7 @@ import 'package:chunked_stream/chunked_stream.dart';
 import 'package:minio/minio.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
-import 'package:unpub/unpub_aws/lib/aws_credentials.dart';
-import 'package:unpub/unpub_aws/lib/s3_file_store.dart';
+import 'package:unpub_aws/unpub_aws.dart';
 
 void main() async {
   group('test s3 file store', () {
@@ -106,139 +105,13 @@ const testPackageData = [
 ];
 
 // This is a tar.gz file with a `hello.txt` file that has `hello` written inside
-const testPackageData2 = [
-  31,
-  139,
-  8,
-  0,
-  4,
-  81,
-  79,
-  97,
-  0,
-  3,
-  237,
-  207,
-  73,
-  10,
-  128,
-  48,
-  12,
-  5,
-  80,
-  215,
-  158,
-  34,
-  39,
-  144,
-  84,
-  99,
-  123,
-  30,
-  209,
-  138,
-  139,
-  98,
-  193,
-  214,
-  233,
-  246,
-  142,
-  116,
-  37,
-  238,
-  84,
-  132,
-  188,
-  205,
-  39,
-  16,
-  194,
-  79,
-  163,
-  141,
-  177,
-  137,
-  159,
-  124,
-  244,
-  28,
-  68,
-  148,
-  68,
-  176,
-  165,
-  82,
-  98,
-  79,
-  76,
-  143,
-  249,
-  36,
-  65,
-  144,
-  72,
-  51,
-  153,
-  99,
-  38,
-  17,
-  80,
-  144,
-  162,
-  60,
-  2,
-  124,
-  176,
-  83,
-  208,
-  59,
-  95,
-  116,
-  107,
-  149,
-  210,
-  26,
-  237,
-  188,
-  30,
-  116,
-  235,
-  108,
-  123,
-  177,
-  183,
-  174,
-  213,
-  245,
-  205,
-  157,
-  227,
-  17,
-  8,
-  249,
-  19,
-  141,
-  29,
-  171,
-  57,
-  254,
-  186,
-  5,
-  99,
-  140,
-  177,
-  183,
-  45,
-  78,
-  193,
-  149,
-  248,
-  0,
-  8,
-  0,
-  0
-];
+const testPackageData2 = [31,139,8,0,4,81,79,97,0,3,237,207,73,10,128,48,12,5,
+  80,215,158,34,39,144,84,99,123,30,209,138,139,98,193,214,233,246,142,116,37,
+  238,84,132,188,205,39,16,194,79,163,141,177,137,159,124,244,28,68,148,68,176,
+  165,82,98,79,76,143,249,36,65,144,72,51,153,99,38,17,80,144,162,60,2,124,176,
+  83,208,59,95,116,107,149,210,26,237,188,30,116,235,108,123,177,183,174,213,
+  245,205,157,227,17,8,249,19,141,29,171,57,254,186,5,99,140,177,183,45,78,193,
+  149,248,0,8,0,0];
 
 String Function(String, String) newFilePathFunc() {
   return (String package, String version) {
